implemented takeDamage for all NPCs, added AbstracNPC

+++ CrossPlatformDesktopProject/Enemy_NPC/INPC.cs
@@ -5,6 +5,6 @@ namespace Sprint2
     {
         float DamageAmount { get; }
 
-        void TakeDamage(int damageAmount);
+        void TakeDamage(float damageAmount);
     }
 }
